modified delete coupon on couponService

+++ backend/src/coupon/coupon.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { createCouponDto } from './dto/create-coupon.dto';
 
@@ -13,6 +17,12 @@ export class CouponService {
   }
 
   async delete(id: string) {
+    if (!isNaN(parseInt(id)))
+      throw new BadRequestException('ID format  not found');
+    const couponCheck = await this.prisma.coupon.findUnique({
+      where: { id: parseInt(id) },
+    });
+    if (!couponCheck) throw new NotFoundException('coupon not found');
     const coupon = await this.prisma.coupon.delete({
       where: { id: parseInt(id) },
     });
